Denon_AVR920.groovy
This is an revised version of the code created by Kristopher Kunicki and Bobby Dobrescu. I had trouble getting the original code to work with my Denon AVR-S920W. So I combined the code from each together to make this one. It will allow you to Power on and off, volume control, mute, and Device selection.